Separate constants from batch_runner module
